requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.17.0
excel-base==1.0.4
idna==3.11
isoweek==1.3.3
python-dateutil==2.9.0.post0
requests==2.33.0
screen==1.0.1
six==1.17.0
sqlparse==0.5.5
TimeConvert==3.0.13
tzlocal==5.3.1
urllib3==2.6.3
xlwt==1.3.0
